log error and close the socket in case SSL_write fails #297
The commit changes the assertion to error-logging & closing the socket, hoping that it would help us determine the root cause of the issue.

My understanding has been that `SSL_write` does not return an error (or partial write) in the pattern that H2O calls the function, but seems like that it has not been the case.