fetch-ocsp-response: Support LibreSSL
Previously, we only used -header when openssl_version starts with
"OpenSSL 1.".  This is because the availability of -header is limited
to that series.  But it turns out that LibreSSL also has this option,
but its version line starts with "LibreSSL ", which does not match our
regexp, and some ocsp responder fails because of lack of Host header
field.  This commit fixes this problem, by allowing sending -header if
version string starts with "LibreSSL ".  LibreSSL's first release tag
is 2.0.0, which is based on OpenSSL 1.0.1h, so it should have -header
option.  Also this commit include fixes to include port part to
ocsp_host.  Technically, Host header field should include port number
(it can omit it if it is a default port for the scheme).