fix crash using libuv, when the socket is closed while there is a pending write
If pending writes exist, libuv calls the write callbacks after the execution of `uv_close` completes.
And the `status` argument passed to the write callback might be zero in case the write succeeded.

Therefore, we would need to retain the pointer to `h2o_socket_t` (which contains `uv_write_req`) and use the variable to detect if the socket has already been closed, and ignore the callback if that is the case.