Suppress casting of function pointer warning
The C standard leaves the casting of function pointers very vague.
The only option is to suppress the warning. This patch adds the
suppression for GCC, others can be added in the future if needed.

The standard says at Section 6.3.2.3, paragraph 8:

> A pointer to a function of one type may be converted to a pointer
> to a function of another type and back again; the result shall
> compare equal to the original pointer. If a converted pointer is
> used to call a function whose type is not compatible with the
> pointed-to type, the behavior is undefined.

Obviously the code works, suppressing the error allows the header to
be included without issue when using GCC.