Introduce a 'proxy.per_socket_pool' option.
When this option is 'ON', socket pooling is done at the frontend
connection level, so that backend connections are only re-used for a
given frontend connection.
We achieve this by adding a `per_socket_pool` member to `h2o_socket_t`.
If the option is set, pooled sockets will be added there and the pool
will be destroyed at the same time as the socket.

This also adds a test that checks that tcp connections are re-using the
same port, there's also a part of the test that checks that connections
over unix sockets are still functional.