Squashed (by kazuho) the changes made in #927
Introduces and uses `h2o_rand` that is implemented as a wrapper for `arc4random` where available.
Also defines `h2o_srand` that should be used on platforms that do not provide `arc4random` and calls the function so that the values returned by `rand()` would differ between the server processes.