Allow to disable libuv even when it is found
It's not currently possible to tell cmake to ignore libuv if it was
found on the system. This add a 'DISABLE_LIBUV' variable (defaults to
OFF), if set, it won't compile any libuv related dependency.

We also move the compilation of unit tests that depend on libuv under
`LIBUV_FOUND`

Finally, this patch replaces `INFO` with `STATUS` as the first argument
of cmake's `MESSAGE` function. See:
https://cmake.org/cmake/help/v3.6/command/message.html