Additional stats
This commit adds duration and jemalloc statistics, both are optional:
jemalloc is enabled at compile time and the duration statistics are
enabled by a 'duration-stats' configuration keyword.
The commit also adds a total number of connections counter.

Duration statistics build on the existing per request duration stats
that are available for logging, each observation is stored in a
structure that allows computing quantiles. When the status page is
queried, values for the minimal, 25th, 50th, 75th and 99th percentile
are output.