Header validation rejects pseudo-headers if sent as literals
Pseudo-headers start with the ':' character (ASCII 0x3a), which causes
the validation logic to kick in and send a protocol error saying that an
invalid character was found in the header.

We fix this by whitelisting all known pseudo-headers (RFC7540 says that
"Endpoints MUST NOT generate pseudo-header fields other than those
defined in this document.")

Unfortunately nghttp2 doesn't allow to test this, because it sends an
indexed header instead of a string literal, which browsers (with the
exception of Safari) apparently do for all non '/' requests.