Fix assertion failure: `sock->super._cb.write != ((void *)0)' failed`.
When a socket was polled for both read and write with DONT_WRITE flag and if a read event occurred, the write-side of the callback was called as well. This resulted in the reset of `_cb.write` but the epoll requests never get updated if the application continues to poll for read (which is the case of HTTP/2 impl.), that eventually ends up in the assertion when the socket really becomes writable.

This PR fixes the issue as well as limiting the call to `getsockopt(SO_ERROR)` to only when establishing the connection.

Fixes #966