Don't stop parsing headers on invalid chars
Since we chose to send a soft error on invalid characters, we can't stop
parsing the headers. In particular, we need to record the errors, in
case the client sends the badly formed header again, but as an index in
the dynamic table.
Make sure that `err_desc` is set to NULL in
`t/00unit/lib/http2/hpack.c::test_request`.