Add an option to prevent emiting x-forwarded-* headers
There are setups where it might not be desirable to add the headers at
all (for example if H2O is not terminating the TLS connection, it will
set `x-forwarded-for:localhost`).

We add a `proxy.emit-x-forwarded-headers` global config option that
instructs H2O not to add or append `x-forwarded-for` and
`x-forwarded-proto`. This option is `OFF` by default.