Broken http/1.1 chunked body is silently relayed
When a `transfer-encoding:chunked` transfer is interrupted in the
backend by a connection close, H2O currently just closes the frontend
stream.
This makes it not possible for the client to know that it's receiving
truncated data, whereas it would be possible over plain HTTP/1.1, since
the last chunk would be missing - Chrome and curl both report errors on
truncated HTTP/1.1 chunked transfers, FF and Safari do not, as far as
I can tell -.
This adds a test that checks that a broken `transfer-encoding:chunked`
transfer is translated into a `RST_STREAM` over H2.