Track stream errors
We introduce `h2o_stream_send_state` which tracks the state of the
stream: finalized, in progress, or error.
We make `h2o_ostream_pull_cb` return this type and `h2o_send` pass it as
an argument. This way protocol implementations can then chose the
appropriate behavior regarding the downstream response.
In particular, we address the case of H2 in `finalostream_send` by
sending `RST_STREAM` if the stream had an error.