One byte connection hung
When receiving a window update message, mark the stream as active if the
stream is about to close. This allows the http2 subsystem to properly
close the stream.

The stream could otherwise become stuck if it stopped writing because
there was no space available in the stream, since no other activity
would wake it up.