Don't send GOAWAY if a PRIORITY frame is sent before the HEADERS
The RFC states that 'PRIORITY can be sent and received in any stream
state'. We correctly accept the `PRIORITY` frame on an `IDLE` stream,
however the code in `handle_headers_frame` would then send a `GOAWAY`
frame on a subsequent `HEADERS` frame.

This happened because the existing code would expect `stream` to only
exist if the `HEADERS` frame had been received, but that also happens
when a `PRIORITY` frame has been received.

We fix this by having `handle_headers_frame` accept `HEADERS` frames if
`stream` exists and is `IDLE`.

This issue was reported by @Lukasa, including a repro case here:
https://github.com/python-hyper/hyper-h2/issues/302