proxy http/2 to http/1: POST+body is missing a content-length
The RFC encourages senders to send a `content-length:` header, even
when it's zero for POST or PUT when no `transfer-encoding` header is
present.

Failing to do so actually leads to some servers issuing a `411 Length
Required` response over HTTP/1, since it's not possible for them to tell
if they got the full request body.

This ambiguity doesn't exist over HTTP/2, since the protocol provides
framing, so some browsers (Safari and Chrome) omit the `content-length:`
header over HTTP/2. H2O should add the header when translating such
requests to HTTP/1.

This commit has the proxy translation layer add a `content-length:`
header if needed, that is when a POST or PUT method is used and there's
no `transfer-encoding` header.