Remove listeners from epoll on SIGTERM
Closing an fd is not enough to stop receiving epoll events, if the
underlying resource is still alive.

This is the case if `h2o` is run in `master` mode: the listener will
still be alive in Server::Starter and the new `h2o` instance. This
results in epoll events being triggered in the old instance, and the
following message to be displayed:

  `ignoring epoll event (fd:-1,event:1)`

We fix this by unregistering the fd from epoll before closing it.