Ignore PRIORITY frames that reference closed pushed streams
Firefox sends a PRIORITY frame when receiving a PUSH_PROMISE frame.
Because of the racy nature of the protocol, it's very possible that the
PRIORITY frame is received once the stream is closed on H2O side.

This results in us opening a new stream that will be used to account
against `globalconf->http2.max_streams_for_priority`. Eventually, this
can lead to sending a GOAWAY frame with 'ENHANCE_YOUR_CALM'.

This commit causes H2O to ignore PRIORITY frames that reference:
- pushed streams that are closed or idle
- pulled streams that are closed

I've tested the fix by using the following handler (and a dummy js file)
and verified that we hit the GOAWAY path without the fix, and not anymore
with the fix:

```
        mruby.handler: |
          Proc.new do |env|
            case env["PATH_INFO"]
            when "/"
              push_paths = []
              (1..100).each { |i| push_paths << "/index.js?#{i}" }
              [399, push_paths.empty? ? {} : {"link" =>
ush_paths.map{|p| "<#{p}>; rel=preload"}.join(",")}, []]
            else
              [399, {}, []]
            end
          end
```