tests are failing because of address re-use
A common case for test failures is due how tests are using
`empty_port()` to obtain addresses to bind to. This is inherently racy,
because of a TOCTOU race: we first check if the port is not used, then
bind it.

This implements a workaround by providing two functions:
`safe_empty_port()` and `safe_empty_port_release()`. When the
SAFE_EMPTY_PORT_DIR environment variable is set, these functions use a
file as a lock in that directory, so that another sibling process
doesn't obtain the same port. When the port is allocated, we create an
empty file with the port name in that directory, and other tests will
keep retrying to obtain a port if the file exists.