Don't bump `conn->pull_stream_ids.max_open` for PRIORITY frames
`PRIORITY` frames can be sent for idle streams, and shouldn't affect the
number of open streams. When registering a stream, determine whether the
registration comes from a `PRIORITY` frame or not, and if it is, don't
affect the `max_open` stream id.

Not doing so resulted in the server closing the connection with a
`PROTOCOL_ERROR` when a `PRIORITY` frame for stream id 3 was followed by
a `HEADER` frame for stream id 1.

@Kriechi provided a minimal repro case
here: https://github.com/mitmproxy/mitmproxy/issues/1824#issuecomment-265709060

I've proposed a PR for h2spec to cover the case:
https://github.com/summerwind/h2spec/pull/67/files