Pushes: support a x-http2-push-only attribute
When this is set, H2O will interpret the `rel=preload` link as a header
we only want to use for H2 pushes, rather than something we also want to
push to the client. So we'll be removing that attribute from the link
header. If there's only one link, we remove the full header.

This is change is there to allow HTTP/1 clients to express HTTP/2 pushes
in a way that doesn't affect the response sent to the client.