[proxy] streaming request body
This change introduces a new possibility for H2O handlers: beeing
passed the request body chunk by chunk. In order to do so, we add a
`on_req_body` function pointer to `h2o_req_t`, that is meant to be
called as data is received. We modify `lib/http2` to use this when the
handler is stream capable (only the proxy handler has this capability).

We also modify `lib/common/http1client.c` to support streaming the
body of the request (examples/libh2o/http1client.c was modified to
desmostrate use this streaming interface).