Make sure all threads have started before starting to serve requests
It's possible that not all threads have started by the time other
threads start to serve requests.

This is in particular problematic with the status handler: the
`on_req_json` function relies on `on_context_init` being called for all
threads, but in fact the call of `h2o_multithread_register_receiver`
might be racing against the value of `self->receivers.size`.

This results in more threads being called than
`collector->num_remaining_threads_atomic`. A consequence of this can be
that either `send_response` is not called (and the fd of the request
is leaked), or on `final()` being called twice (leading to a possible
double free).

This commit solves the issue by introducing a synchronization point
in the initialization, in order to make sure that all contexts for all
threads are initialized before starting to serve requests.