undefined behavior fixes: h2o_memcpy
Invoking `memcpy(dst, NULL, 0)` is a common pattern in H2O.
Unfortunately, the C standard says passing a NULL pointer to either dst
or src leads to undefined behavior. The glibc follows this by using the
`nonnull` attribute to annottate both pointers.

This means that a optimizing compiler could take the opportunity to
assert that `src` is *never* NULL, and remove valid NULL checks from the
code.

We introduce a `h2o_memcpy` variant that special cases `src` being NULL
to avoid this issue.