Dynamic backends
This commit adds infrastructure to make it possible to dynamically chose
a backend destination on a per request basis. We then take advantage of
this by allowing `proxy.reverse.url` to take a mapping, in which case we
can add customizable routing policies. This commit adds the `header`
policy, if this is set, H2O will look for a given header name in order
to chose the backend destination.

Having dynamic backends also supposes that backends can defined on their
own right, indenpendently of `proxy.reverse.url`.  To that end, this
commit defines a `backend` directive that allows to declare a backend to
H2O, which can then be used for routing.