Increase the idle timeout.
I have seen test failures multiple times, caused by the `idle-timeout` (2 seconds) being triggered. The failures happen on the HTTPS + ASAN test only.

Considering the fact that the failures are happening on the most CPU-intensive variant, the idle timeout is being increased in hope that doing so would fix the issue.