Fix deadlock in setup_bio()
setup_bio() uses a mutex to guard the initialization of bio_methods,
which is never released. As a consequence, if two threads find
bio_methods to be NULL at the same time, or if one of them is in the
critical section while the other one tries to acquire the mutex, one
thread will go on and initialize bio_methods and the other one will wait
for the mutex to be released indefinitely.

Fix this by releasing the mutex after bio_methods is initialized.