ssl: avoid spurious ticket renewals
H2O currently signals OpenSSL to unconditionally renew a ticket if it
was not encrypted using the latest key available. While this generally
works, it assumes that the latest key is always valid for encryption,
which however may not be the case. When encryption keys are rotated,
the newly-generated key's `not_before` is set to 60 seconds in the
future, so the most recent key is not yet valid for a whole minute. This
in turn leads to unnecessary ticket renewals, where a ticket will be
re-issued using the same (active) encryption key.

We fix this by checking that the most recent key can be used before
requesting renewal.