I/O errors with multiple 1xx responses
The issue happens if the responses are received in the same buffer. We
succesfully parse the first response, but we don't look at the buffer to
see if there might be another one to parse.
Sending multiple 103 responses is explicitely permitted by RFC 8297

We tweak the existing push test to send the same push twice, so that the
code path is exercised.