Disable renegotiation on client side SSL contexts
The current behavior of the ssl stack is to close the socket if a read
triggers WANTS_WRITE, meaning that the peer starts a renegotiation.

This PR modifies the SSL_CTX in order to advise the peer that we're not
doing renegotiations. If the peer still attemps a renegotiation, we'll
still close the connection.