Add CMake option to disable filebacked-buffer
Add CMake option "FILEBACKED_BUFFER", allowing the user to allow (the
default) or disallow h2o of making use of a file-backed buffer with
mmap.

The default is "ON", which leaves the current behavior as it is: under
specific conditions, a buffer might be migrated to an mmapped area, in
the filesystem.  If set to "OFF", a request that might depend on such
large buffer will be denied as an mmapped area and thus will only reside
in main memory.

You should disallow it (-DFILEBACKED_BUFFER=off) only if you never want
h2o to create a file-based mmapped buffer, and your system has plenty of
RAM.