add argument to setup_ecc_key
The new signature is setup_ecc_key(SSL_CTX, const char *). The additional argument can have one of the following values:

 - "auto": Follow the old behaviour to enable the default list of ECC curves if possible, which may be not-so-secure (see #713).
 - A list of curve names or NIDs, seperated by a colon (e.g.: "P-256:brainpoolP512r"). A fairly recent version of Open/LibreSSL is required. With OpenSSL 1.1.0 group names like "ffdhe4096" may work for TLS 1.3, too.
 - A short name of a curve (e.g. "brainpoolP512r"). This will work with any OpenSSL version supporting ECDH.

If the linked Open/LibreSSL does not support ECDHE, the silently does nothing. This may change in the future. If anything fails, the function returns but h2o won't crash.

The behaviour is more or less identic to nginx'. In the current state the curve settings are compile time only.