`h2o_http2_conn_unregister_stream` assert on shutdown
`h2o_http2_conn_unregister_stream` wants `_refs.link` set if
the state is pending.  However, when we sent `GOAWAY` (for
draining), we don't bother linking said struct member in
`execute_or_enqueue_request_core`. This fixes the issue by not switching
the stream to `REQ_PENDING` if we're going to reset it.