file: use buffer_try_reserve for dir listing buffer
While this change might be too defensive, I was able to trigger
an mmap call by preparing a large directory on the APFS file system,
which is comparable to other file systems in this particular context.
I admit that my test case is unorthodox but at the same time, I can't
say that it's impossible for an innocent user to have a big directory.