socket: use h2o_buffer_try_reserve where applicable
This change ensures that the behavior on buffer allocation failure
remains the same at the socket layer. Even though the allocation
amount is fixed to 4096-bytes inside the event loop backends, calls
to h2o_buffer_reserve was edited to use h2o_buffer_try_reserve to
allow the event library to handle what to do on allocation failure.