mruby: use h2o_buffer_try_reserve where applicable
Buffer allocation in send_chunk_method() need to be replaced with
h2o_buffer_try_reserve, because the amount of memory allocated grows
proportionally to the size of the output. On the other hand, the
buffer allocation in post_error() can remain as-is, because the
allocation size is determined by short constant strings.