Add support for packet timestamping on Linux.
This change allows the user to specify the option "packet-timestamping" to turn
on packet timestamping on supported platforms (this change implements this for
Linux).

The user may also specify a "packet-timestamping-sample-rate" to determine how
often to read the packet timestamping information and compute the average.

The data is exported in nanoseconds per thread via "duration-stats" (if enabled).