Expand "openssl_backport.h" to include BIO_meth_free() and BIO_meth_set_*().
These macros aren't used by h2o, but this header file is the first
search engine hit for someone looking to polyfill OpenSSL 1.0.2 so
that a single program can work portably against either 1.0.2 or 1.1.0.
Therefore, it is helpful to the wider community if this header file
contains the *full* polyfill for `BIO_METHOD` setters added in 1.1.0.

OpenSSL 1.1.0 also adds getters such as `BIO_meth_get_ctrl(bm)`.
I have no need for them myself, but the same logic applies: this
useful header file should be a full polyfill for the benefit of
the wider community.