Bring back uint8_t for frame type with workaround
This patch again defines the frame type argument for the HTTP/2
unknown frame probe as uint8_t, which is the correct type for
frame type.

As of now, bpftrace is known to grab 64-bit integer from the memory
regardless of the datatype declared in h2o-probes.d, which sometimes
grabs garbage data in a adjacent memory address.
Thus this patch implements a workaround to clear the upper 56 bits of
the argument.