Add the ability to specify address family for proxy connections
When the family key is used, the address family for the backend specification
can be controlled. Legal values for are inet and inet6. If nothing is specified,
an address family of PF_UNSPEC is used, which will result in both IP and
IPv6 (if present) being used.

Example:

proxy.reverse.url:
  - url: http://some.host.com:8080
    family: inet