Use closefrom(2) if available in the OS
Currently, when the neverbleed process is forked we query _SC_OPEN_MAX and
iterate through every file descriptor above STDERR (skipping listen and
close notify file descriptors) and call the close(2) syscall on the fd.
Each syscall is a context switch, and depending on the audit/debug state of
the system during this time, roughly 231,000 syscalls will be executed.
On security audited, traced or otherwise hooked (kprobes, ebpf, perf)
processes, this will generate 231,000 records being committed to disk.

This change introduces an optimization which determines the highest file
descriptor value currently in the process, uses the standard close(2) syscall
for anything less than (excluding stdout, stderr, and stdin) then executes
closefrom on the highest fd plus one. This substantially decreases the number
of context switches during the bootstrap phase of the neverbleed process (from
231000 to roughly 9. This change also includes the necessary cmake changes to
detect the functionality. Tested and verified on FreeBSD:

 [..]
 52544: close(3)                                  = 0 (0x0)
 52544: close(4)                                  = 0 (0x0)
 52544: close(5)                                  = 0 (0x0)
 52544: close(6)                                  = 0 (0x0)
 52544: close(7)                                  = 0 (0x0)
 52544: close(8)                                  = 0 (0x0)
 52544: close(9)                                  = 0 (0x0)
 52544: close(11)                                 ERR#9 'Bad file descriptor'
 52544: closefrom(13)                             = 0 (0x0)
 [..]