Use operating system to check for closefrom(2)
This change set is going into the neverbleed repository and it
doesn't have cmake. Remove cmake code to check for the feature
and use the operating system as an indication as to whether or
not this feature is supported.

NOTE: I am pretty sure Solaris has this too, but I need to find
a system I can do the build+test on before I add it to the list.