Add workaround for CIFuzz build bug
CIFuzz currently does not support fuzzing of forks. See https://github.com/google/oss-fuzz/issues/3731. This changes the configuration so that it CIFuzz does not run on forks, i.e. does not run on PRs from forked repos.

This change should be rolled back when the CIFuzz bug is fixed.