Conditionally use libc wrapper based on sandbox
The current seccomp policy for FastCGI permits open(2) in the outer
sandbox for Linux.  FreeBSD's sandbox (capsicum) such operations are
not permitted and will need to go through the wrappers. We should
modifify this code to function in Linux too so we can remove the
open() requirements for the seccomp mode of operation