Don't set NODELAY for non-inet sockets.
strace allows you to trace a process printing only failed system calls (which
is extremely useful for debugging). If lots of connections are generated
calling `setsockopt` on a socket that doesn't support NODELAY can fill your
strace logs with errors and make it harder to debug syscall failures you are
interested in.

This change avoids NODELAY for UNIX sockets. In addition to easier `strace`
debugging, we can also remove an unncessary system call.