Do not try to forward to self.
When h2o receives an Initial packet with corrupt SCID, that packet would always be passed to the `forward_packets` callback, as the packet cannot be processed and therefore there would not be connection state. The forward destination, which is determined by the DCID field of that Initial packet can point to the running thread itself (in practice, this happens when a client sends an Initial packet in response to server-sent Initial; consider HelloRetryRequest).

As `forward_packets` callback cannot be called with the destination being set to the running thread, we need a guard.