Add test case for 2nd-flight Initial with corrupted SCID
There was a known issue in h2o that when it received non-decryptable
second-flight Initial, while decrypted DCID pointed to the receiving
node and thread, h2o crashed due to an assertion failure.

This test case reproduces such a scenario by injecting a second-
flight Initial with corrupted SCID (which makes payload decryption
fail and h2o think it should be forwarded to a node pointed by
decrypted DCID).