`strict-match: ON` in a host entry makes it ineligible for use as the fallback host entry.
Requests without :authority (or a Host header field) and requests that do not
match any of the configured host entries will be routed to the first host entry
that does not have the strict-match flag.  If all entries have the flag, then
a request that does not match any of them will get a 404 "not found".

This allows to list specific entries first before a wildcard entry and
still to route requests without Host (or :authority) to the wildcard entry.

Example:

```
hosts:
  "www.example.com":
    strict-match: ON
    paths:
      /:
        file.dir: examples/doc_root1
  "*.example.com":
    strict-match: ON
    paths:
      /:
        file.dir: examples/doc_root2
  "*":
    paths:
      /:
        file.dir: examples/doc_root3
```

In this example, requests without a Host header will be routed to the third
host entry.

Without a way to mark host entries as ineligible to be the default virtual host,
requests without a Host header can only be routed to the first host entry in
the configuration, and that is not always the appropriate destination for those
requests.