Configure mruby with the same C compiler as the one that cmake uses.
`deps/mruby/tasks/toolchains/clang.rake` looks for `CC`, `CXX` and `LD`, with
hard-coded defaults for each when unset.  In particular, `CC` defaults to
`clang` (to be found on `$PATH`) when `MRUBY_TOOLCHAIN` is `clang`.

However, the build may have configured `cmake` with a different executable for
the compiler, for example `clang-10`.

Therefore we now supply this information to `ruby minirake` through the
environment, populated from the values that cmake has been using for compiling.

There are no C++ files in `deps/mruby` and therefore we can use the C compiler
for linking rather than the C++ compiler.