Add a `group` attribute under the `listen` directive to accompany the `owner` attribute.
When `owner` is specified and `group` is not, then the listening socket will be
created with the specified user's primary group, as before.  This commit does
not change the default behavior.

When both `owner` and `group` are specified, then `owner` specifies the user id
for the listening socket and `group` specifies the group id for the listening
socket.

It is a configuration error to specify `group` without specifying `owner`.