Add a `max_conn_backlog` field to the configuration structure
This allows H2O to perform load-shedding in the kernel by lowering the
number of pending connections on a given listener socket from
`H2O_SOMAXCONN` to a more reasonable value. The default value is kept as
`H2O_SOMAXCONN`, and the line `max_conn_backlog: NUMBER` can be set in a
configuration file to lower it away from that.