fix: do while not required here
do { ... } while(0) is used as a macro contruct to wrap multi-line statements so
they appear as a single statement.

It is not required to do this here as we're in the macro's context.

Moreover, this prevents the use of `continue` and `break` within block
as it get's applied to the do-while and not the for loop above as
intented.

Ref: https://wiki.sei.cmu.edu/confluence/display/c/PRE10-C.+Wrap+multistatement+macros+in+a+do-while+loop