Add config command `proxy.http2.force-cleartext`.
`proxy.http2.force-cleartext: ON` together with `proxy.http2.ratio: 100`
and with an `http` URL in `proxy.reverse.url` instructs the proxy to use
the cleartext HTTP/2 protocol without discovery or negotiation.

The h2o configuration will be rejected if `proxy.http2.force-cleartext`
is set to `ON` yet `proxy.http2.ratio` is not set to `100`,

This configuration command has no effect if the URL in `proxy.reverse.url` is
an `https` URL because over TLS the ALPN negotiation is mandatory for HTTP/2.