Separate `startup_sync_barrier` into two barriers, the first only for run_loop threads.
The run_loop-thread-only barrier is named `startup_sync_barrier_init`.  It
gates one-time initialization code for run_loop threads.  Only run_loop threads
synchronize on this barrier.

The second barrier is named `startup_sync_barrier_post`.  It does not gate any
one-time initialization code for run_loop threads.  It therefore does not
matter whether the last thread to cross the barrier is a run_loop thread or not.

Before this commit, when the last thread to cross the barrier was the thread
that runs `register_session_tickets`, then the initialization code for run_loop
that the barrier was gating did not run.