fix integer arithmetic error
(int-int)/size_t is mixed operation where the resulting int
(negative if # connections < threshold) get's converted to a size_t,
resulting in a large number, > 0, therefor was always culling connections.

This does not seem to be an issue when the division is by 1 (number of threads)
which the tests are using.

Related:
https://en.cppreference.com/w/c/language/conversion
"Usual arithmetic conversions" 4)
- Else, the operands have different signedness:
    - If the unsigned type has conversion rank greater than or equal to
      the rank of the signed type, then the operand with the signed type
      is implicitly converted to the unsigned type.