Do not try to connect to the same address repeatedly in the connect handler.
When the connect function returns with an error immediately,
do not carry over the already tried server_address into the
next connect attempt, because that can cause an infinite loop.

The test case demonstrates the problem using `255.255.255.255`.
A non-blocking connect to this address fails immediately on the
operating systems that I have access to (Linux, MacOS and OpenBSD),
albeit with a different errno on each (ENETUNREACH, EAFNOSUPPORT
and EINVAL respectively).