Potential assert when the timeout coincides with a write
The connect handler currently keeps the the io timeout armed after a
write has been issues to front end handlers. This is a problem because
it's possible that the timeout fires when there's a write is already in
filght, which is not the contract that handlers have with frontends
(only one outstanding h2o_send can be inflight).

We fix this by simply disabling the timeout when a write is in flight,
relying on the frontend idle timeout instead in these cases. The timeout
is re-armed in the 'proceed' callbacks, as soon as the frontend is done
with the write.