Fix a crash in the CONNECT handler after timeout with a write to client still pending.
If a write to the client is in progress when `proxy.timeout.io` expires, and
there is data in the send buffer from the client to the origin, then
`close_readwrite` will run the first time, close the origin socket, set
`self->sock` to NULL, and it will schedule itself to run again.

When `close_readwrite` runs the second time, `self->sock` is already NULL,
`close_socket` will not be called and `send_inflight` will not be determined
correctly.  As a result of that, `close_readwrite` will call `h2o_send` even
though a write to the client is still pending.

The test `t/50connect-inflight.t` replicates the problem.

The fix in this commit is to determine `send_inflight` from the detached write
buffer when `self->sock` is NULL.