HTTP/3 server: send a reset when do_send errors out
The current behavior of http3/server.c::do_send() is to close the stream,
there is thus no way for the client to tell that there might be an issue
with the response.

The code as proposed in this commit does present the issue noted
in https://github.com/quicwg/base-drafts/issues/3300: a re-ordering
of packets might result in a worse end user experience compared to a
TCP based protocol.