[http2 server] when possible send an END_STREAM flag with HEADERS
We currently do not send a HEADERS frame with an END_STREAM, relying
instead on the ability to send the flag along with a zero sized DATA
frame. This causes interoparibility issues with some clients that treat
the presence of a DATA frame (empty or not) as semantically different.

This PR makes it so that we set the flag when generating the HEADERS
frame, if the information is available at the time (the generator
indicated that this is the end of the stream).