Fix proxy_idle_time to exclude client latency
This change measures proxy_idle_time as the time from when we create the
backend client until we receive the first byte from the backend.

Prior to this patch we measured the time from when we received the first
byte from the client until we received the first byte from the backend
which inadvertently included network and client latency resulting in
inflated proxy_idle_times.