http/3 quic: fix inter node forwarding for mismatched address families
Current inter-node forwarding doesn't work for for mismatching address
families: if the internode address is ipv4, it cannot process forwarded
ipv6 packets and vice versa.

This fix instruments the listener creation logic by marking which
listeners are each other's 'siblings', meaning they come from the same
configuration directive, which resolved to one ipv4 listener on one hand
and ipv6 on the other.

The forwarding logic then uses that sibling fd in order to reach the
worker thread with the context matching the address family.