HTTP/3: Reject leading and/or trailing space in Huffman-coded fields
h2o_hpack_decode_huffman() didn't check for the absence of leading
and/or trailing whitespace in field values.  The HTTP/2 code did check
for this, but the HTTP/3 code didn’t.  Fix this by having
h2o_hpack_decode_huffman() do the check itself, and remove the check
from its caller as it's now redundant.  Both HTTP/2 and HTTP/3 already
checked for leading and/or trailing spaces in non-Huffman-encoded
fields via h2o_hpack_validate_header_value(), but this is not called for
Huffman-encoded headers as the checks are instead done during Huffman
decoding.

Fixes: #3471